move Future-related functionality into subspec
