+++ Hyperspace.podspec
@@ -32,15 +32,23 @@ Unlike other networking libraries, the goal of Hyperspace is to keep things simp
   s.ios.deployment_target = '8.0'
   s.watchos.deployment_target = '2.0'
   s.tvos.deployment_target = '9.0'
+  s.default_subspec = 'Core'
 
-  s.source_files = 'Sources/Hyperspace/**/*'
+  s.subspec 'Core' do |core|
+  core.source_files = 'Sources/Hyperspace/**/*'
 
-  # s.resource_bundles = {
+  # core.resource_bundles = {
   #   'Hyperspace' => ['Hyperspace/Assets/*.png']
   # }
 
-  # s.public_header_files = 'Pod/Classes/**/*.h'
-  # s.frameworks = 'UIKit', 'MapKit'
-  s.dependency 'Result', '~> 4.0'
-  s.dependency 'BrightFutures', '~> 7.0.0'
+  # core.public_header_files = 'Pod/Classes/**/*.h'
+  # core.frameworks = 'UIKit', 'MapKit'
+  core.dependency 'Result', '~> 4.0'
+  end
+
+  s.subspec 'Futures' do |futures|
+  futures.dependency 'BrightFutures', '~> 7.0.0'
+  futures.dependency 'Hyperspace/Core'
+  futures.source_files = 'Sources/Futures/**/*'
+  end
 end
